requirement.txt
langchain==0.3.20
langchain-community==0.3.19
python-dotenv==1.0.1
langchain-google-genai==2.0.11
duckduckgo-search==7.5.0
selenium==4.29.0
pandas==2.2.3
langchain-cohere==0.4.3
python-dotenv==1.0.1
langchain-ollama==0.3.0
langchain-groq==0.3.1
streamlit==1.43.2
numpy==2.2.4
webdriver-manager